requirements.txt
dill==0.3.4
transformers==4.26.1
faiss-gpu==1.6.3
wandb==0.13.10
torch==1.0.1
pytorch-metric-learning==1.3.2
pandas==0.23.4
Pillow==9.1.1
tqdm==4.26.0
torchvision -c pytorch
requests==2.27.1
urllib3==1.26.9

